requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
MouseInfo==0.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==1.0.0
pytweening==1.0.4
requests==2.31.0
tabulate==0.9.0
urllib3==2.0.3
